Completely centralise DashboardGroup and Project models
