+++ packages/database/src/modelClasses/Project.js
@@ -23,4 +23,21 @@ export class ProjectModel extends DatabaseModel {
   get DatabaseTypeClass() {
     return ProjectType;
   }
+
+  async getAllProjectDetails() {
+    return this.database.executeSql(`
+      select p.id, p.code,
+            to_json(sub.child_id) AS entity_ids,
+            e."name", p.description,
+            p.sort_order, p.user_groups,
+            p.entity_id, p.image_url,
+            p.logo_url, p.dashboard_group_name,
+            p.default_measure
+      from project p
+        left join entity e
+          on p.entity_id = e.id
+        left join (select parent_id, json_agg(child_id) as child_id from entity_relation er group by parent_id) sub
+          on p.entity_id = sub.parent_id
+    `);
+  }
 }
