+++ packages/database/src/modelClasses/index.js
@@ -80,13 +80,11 @@ export { AlertModel } from './Alert';
 export { CommentModel } from './Comment';
 export { CountryModel } from './Country';
 export { DataSourceModel } from './DataSource';
-export { DashboardGroupModel } from './DashboardGroup';
 export { EntityModel, EntityType } from './Entity';
 export { FacilityModel } from './Facility';
 export { GeographicalAreaModel } from './GeographicalArea';
 export { MeditrakDeviceModel } from './MeditrakDevice';
 export { PermissionGroupModel } from './PermissionGroup';
-export { ProjectModel } from './Project';
 export { SurveyScreenComponentModel } from './SurveyScreenComponent';
 export { SurveyScreenModel } from './SurveyScreen';
 export { UserEntityPermissionModel } from './UserEntityPermission';
