+++ packages/web-config-server/src/models/index.js
@@ -3,12 +3,8 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-import { ProjectModel } from './Project';
-import { DashboardGroupModel } from './DashboardGroup';
 import { EntityModel } from './Entity';
 
 export const modelClasses = {
-  DashboardGroup: DashboardGroupModel,
   Entity: EntityModel,
-  Project: ProjectModel,
 };
